pytest(conftest): move workdir bindir and cmake_file_api
To a centralized place.
