+++ tests/python/.pytest.ini
@@ -1,6 +1,7 @@
 [pytest]
 addopts =
     --strict-config
+    --strict-markers
     -vvv
     -s
     --log-cli-level=info
